Use Location detail for return type
