--- src/main/java/org/opensrp/domain/LocationBean.java
@@ -1,61 +1,0 @@
-/**
- * 
- */
-package org.opensrp.domain;
-
-/**
- * @author Samuel Githengi created on 05/06/20
- */
-public class LocationBean {
-	
-	private Long id;
-	
-	private String identifier;
-	
-	private String name;
-	
-	private boolean voided;
-	
-	private String tags;
-	
-	public Long getId() {
-		return id;
-	}
-	
-	public void setId(Long id) {
-		this.id = id;
-	}
-	
-	public String getIdentifier() {
-		return identifier;
-	}
-	
-	public void setIdentifier(String identifier) {
-		this.identifier = identifier;
-	}
-	
-	public String getName() {
-		return name;
-	}
-	
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	public boolean isVoided() {
-		return voided;
-	}
-	
-	public void setVoided(boolean voided) {
-		this.voided = voided;
-	}
-	
-	public String getTags() {
-		return tags;
-	}
-	
-	public void setTags(String tags) {
-		this.tags = tags;
-	}
-	
-}
