+++ src/main/java/org/opensrp/domain/LocationDetail.java
@@ -3,25 +3,57 @@ package org.opensrp.domain;
 import java.io.Serializable;
 
 public class LocationDetail implements Serializable {
-    private static final long serialVersionUID = 7360003982578282029L;
-
-    private String identifier;
-
-    private String name;
-
-    public String getIdentifier() {
-        return identifier;
-    }
-
-    public void setIdentifier(String identifier) {
-        this.identifier = identifier;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
+	
+	private static final long serialVersionUID = 7360003982578282029L;
+	
+	private Long id;
+	
+	private String identifier;
+	
+	private String name;
+	
+	private boolean voided;
+	
+	private String tags;
+	
+	public String getIdentifier() {
+		return identifier;
+	}
+	
+	public void setIdentifier(String identifier) {
+		this.identifier = identifier;
+	}
+	
+	public String getName() {
+		return name;
+	}
+	
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+	public Long getId() {
+		return id;
+	}
+	
+	public void setId(Long id) {
+		this.id = id;
+	}
+	
+	public boolean isVoided() {
+		return voided;
+	}
+	
+	public void setVoided(boolean voided) {
+		this.voided = voided;
+	}
+	
+	public String getTags() {
+		return tags;
+	}
+	
+	public void setTags(String tags) {
+		this.tags = tags;
+	}
+	
 }
